requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
groq==0.4.2
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
reportlab==4.0.7
Pillow==10.1.0
gunicorn==21.2.0